% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_check_dates.R
\name{statx_check_dates}
\alias{statx_check_dates}
\title{Return a list of dates from schema endpoint}
\usage{
statx_check_dates()
}
\value{
\code{list}. A list of the available dates for the given dataset.
}
\description{
\code{statx_check_dates} returns a list from \href{https://stat-xplore.dwp.gov.uk/webapi/jsf/login.xhtml}{Stat-Xplore} schema endpoint.
}
\details{
\code{statx_check_dates} returns a list from \href{https://stat-xplore.dwp.gov.uk/webapi/jsf/login.xhtml}{Stat-Xplore} schema endpoint.
}
