% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_dataset.R
\name{statx_dataset}
\alias{statx_dataset}
\title{Download a dataset}
\usage{
statx_dataset(requestBody)
}
\arguments{
\item{requestBody}{\code{string}. A valid JSON request body. See the \href{https://stat-xplore.dwp.gov.uk/webapi/online-help/Open-Data-API-Table.html}{Stat-Xplore help}}
}
\value{
\code{tibble}.
When invalid arguments are used returns \code{NULL} with \code{warning}.
}
\description{
\code{statx_dataset} returns a dataset from \href{https://stat-xplore.dwp.gov.uk/webapi/jsf/login.xhtml}{Stat-Xplore}
}
\details{
\code{statx_dataset} returns either a dataset from \href{https://stat-xplore.dwp.gov.uk/webapi/js/login.xhtml}{Stat-Xplore}.
The \code{requestBody} parameter must be passed a valid JSON request body.
}
