% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_request_dates.R
\name{statx_request_dates}
\alias{statx_request_dates}
\title{Return dates in request body}
\usage{
statx_request_dates(request)
}
\arguments{
\item{request}{\code{string}. A valid JSON request body. See the
\href{https://stat-xplore.dwp.gov.uk/webapi/online-help/Open-Data-API-Table.html}{Stat-Xplore help} and \code{statx_add request}
for advice on creating a request body.}
}
\value{
\code{tibble}. A tibble of the data described in the JSON request.
}
\description{
\code{statx_request_dates} returns all dates from a Stat-Xplore request body
}
\details{
\code{statx_request_dates} returns all dates included in a Stat-Xplore json request body.
The \code{request} parameter must be passed a valid JSON request body. Request bodies can be added to the package
by calling \code{statx_add request}, and existing requests can be displayed by calling \code{statx_requests}.
}
