% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_schema_request.R
\name{statx_schema_request}
\alias{statx_schema_request}
\title{Senc a request to the Stat-Xplore schema endpoint}
\usage{
statx_schema_request(endpoint, access_key)
}
\arguments{
\item{endpoint}{\code{string}. The API endpoint to query.}

\item{access_key}{\code{string}. A valid Stat-Xplore access key.}
}
\value{
\code{list}.
}
\description{
\code{statx_schema_request} sends a request to the Stat-Xplore schema endpoint
}
\details{
\code{statx_schema_request} sends a request to the Stat-Xplore schema endpoint.
It is an internal function, called in \code{statx_check_dates}.
The \code{endpoint} parameter must be a valid schema endpoint.
The \code{access_key} parameter a valid Stat-Xplore access key.
}
\keyword{internal}
