% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_set_renviron_key.R
\name{statx_set_renviron_key}
\alias{statx_set_renviron_key}
\title{Add or update .Renviron key}
\usage{
statx_set_renviron_key(key, value)
}
\arguments{
\item{key}{\code{string}. The key to add or update.}

\item{value}{\code{string}. The value to apply to the key.}
}
\value{
\code{logical}.
\code{TRUE} when key is successfully updated.
If \code{FALSE} will return error message.
}
\description{
\code{statx_set_renviron_key} adds or updates a key in .Renviron file
}
\details{
\code{statx_set_renviron_key} adds or updates a key in \code{.Renviron}.
It is an internal function, called in \code{statx_set_api_key} and \code{statx_set_request_folder},
preventing other keys being accidentally altered.
If there is no existing .Renviron in the home directory it is created.
For changes to take effect the R session must be restarted.
The \code{renviron_key} parameter must be passed a name for the key.
The \code{value} parameter must be passed the value to assign to the key.
}
\keyword{internal}
