% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_dataset.R
\name{statx_dataset}
\alias{statx_dataset}
\title{Download a dataset}
\usage{
statx_dataset(request)
}
\arguments{
\item{request}{\code{string}. A valid JSON request body. See the
\href{https://stat-xplore.dwp.gov.uk/webapi/online-help/Open-Data-API-Table.html}{Stat-Xplore help} and \code{statx_add request}
for advice on creating a request body.}
}
\value{
\code{tibble}. A tibble of the data described in the JSON request.
}
\description{
\code{statx_dataset} returns a dataset from \href{https://stat-xplore.dwp.gov.uk/webapi/jsf/login.xhtml}{Stat-Xplore}
}
\details{
\code{statx_dataset} returns a dataset from \href{https://stat-xplore.dwp.gov.uk/webapi/js/login.xhtml}{Stat-Xplore}.
The \code{request} parameter must be passed a valid JSON request body. Request bodies can be added to the package
by calling \code{statx_add request}, and existing requests can be displayed by calling \code{statx_requests}.
}
