% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statx_set_request_folder.R
\name{statx_set_request_folder}
\alias{statx_set_request_folder}
\title{Add or update the 'RequestBodyFolderPath' in .Renviron}
\usage{
statx_set_request_folder(folder_path)
}
\arguments{
\item{folder_path}{\code{string}. An existing folder path, which is where request bodies will be stored and run from.}
}
\value{
\code{logical}.
\code{TRUE} when key is successfully updated.
If \code{FALSE} will return error message.
}
\description{
\code{statx_set_request_folder} adds or updates the 'RequestBodyFolderPath' key in .Renviron file
}
\details{
\code{statx_set_request_folder} adds or updates the 'RequestBodyFolderPath' key in \code{.Renviron} using the internal function \code{statx_set_renviron_key}.
The \code{folder_path} parameter must be passed an existing folder path, which is where request bodies will be stored and run from.
Backslashes must be escaped, e.g. \code{C:\\Documents\\FolderName} must be input as \code{C:\\\\Documents\\\\FolderName}.
}
